WaitOnAFrame should be on ITestOperations
